CC-4259: stopping services causes nearly all services to be terminated
